+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Zpushadmin: Available for admins only by default
 - Carddav: Share carddav with internal on install
 - Support: Export broken
+- Core: update old dependency for spreadsheet export
 
 26-02-2024: 6.8.36
 - Calendar: Fixed changing color for calendars and categories issue #1112